### 2024-11-12 #7452 @editorialbot
[PRE REVIEW]: MFEM/MGIS, a HPC mini-application targeting nonlinear thermo-mechanical simulations of nuclear fuels at mesoscale
**Submitting author:** @thelfer (<a href="http://orcid.org/0000-0003-2460-5816">Thomas Helfer</a>)
**Repository:** https://github.com/thelfer/mfem-mgis
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
