### 2025-01-24 #7452 @jacobmerson
@thelfer @arfon I am willing to review and I have a good amount of overlap with the topic. I have direct experience with MFEM and my background is in nonlinear solid mechanics. Most of my current research is putting together multiscale and multiphysics simulations for exascale fusion power/plasma simulations. I know almost nothing about nuclear fuels, but I assume that's not the main issue you need reviewed for the JOSS article.
