### 2025-01-24 #7452 @thelfer
>  I know almost nothing about nuclear fuels, but I assume that's not the main issue you need reviewed for the JOSS article.

No, it has that we probably more focused on robustness, accuracy and long term support than speed with a focus on multiphysics (with partioned schemes)
