### 2025-01-25 #7452 @arfon
@iammix, @jacobmerson – many thanks for agreeing to review this submission for JOSS! See you and @thelfer over in #7719 where the actual review will take place.
