+++ .github/workflows/cmake.yml
@@ -41,6 +41,7 @@ jobs:
 
     - name: Install Packages
       run: |
+        spack compiler find
         spack external find openmpi mpich libfabric
         spack install mfem+mpi+suite-sparse+mumps@4.7.0
         spack install tfel@master:~python~python_bindings
