+++ .github/workflows/cmake.yml
@@ -43,9 +43,9 @@ jobs:
       run: |
         spack compiler find
         spack external find openmpi mpich libfabric
-        spack install mfem+mpi+suite-sparse+mumps@4.7.0
-        spack install tfel@master:~python~python_bindings
-        spack install mgis@master:+c~fortran~python
+        spack install mfem+mpi+suite-sparse+mumps@4.7.0%gcc@12.3.0
+        spack install tfel@master:~python~python_bindings%gcc@12.3.0
+        spack install mgis@master:+c~fortran~python%gcc@12.3.0
 
     - name: Config MFEM-MGIS [coverage]
       shell: spack-sh {0}    
