+++ .github/workflows/cmake.yml
@@ -53,8 +53,8 @@ jobs:
         spack load mfem 
         cmake -B ${{github.workspace}}/coverage \
               -DCMAKE_CXX_FLAGS_COVERAGE="--coverage -g" \
-              -DCMAKE_BUILD_TYPE=Coverage -DCMAKE_C_COMPILER=gcc \
-              -DCMAKE_CXX_COMPILER=g++ \
+              -DCMAKE_BUILD_TYPE=Coverage -DCMAKE_C_COMPILER=gcc-12 \
+              -DCMAKE_CXX_COMPILER=g++-12 \
               -DMFEM_DIR=`spack location -i mfem` \
               -DHYPRE_DIR=`spack location -i hypre` \
               -DMFrontGenericInterface_DIR=`spack location -i mgis`/share/mgis/cmake \
