### 2025-02-28 thelfer/mfem-mgis#102 @jacobmerson
JOSS: paper comments
Comments paper related to https://github.com/openjournals/joss-reviews/issues/7719

- The paper focused on the integration of MGIS with MFEM and had essentially no reference to nuclear fuels or mesoscale simulations. This isn't necessarily a problem as I think the integration with MGIS is a useful contribution on its own. But, either the title should remove the discussion of nuclear fuels, or the text should provide more relevant background.
- There are many finite element solvers for nonlinear solid mechanics problems. The paper is missing a discussion of how this relates to any of those. E.g., Moose, code-aster, abaqus, ansys, etc.
- Some additional discussion on what is MFront and how that generalizes the integration of material models into MFEM is needed.
- You mention that integration into MFEM will allow you to "efficiently use supercomputers". Have you done any scaling, or demonstration of this?
- The way the first sentence in the statement of need is written (line 33) is misleading. MFEM is a finite element framework and doesn't make any assumptions about the types of constitutive laws that you implement.
- line 46, you need to demonstrate that you can "maintain parallel performance"
- line 48 micromorphique -> micromorphic
- You specifically mention micromorphic and orthotropic material laws. Is there a limitation in the current implementation that prevents other laws that are implemented in MFront? You also mention plasticity in the introduction, is this included?
- I'm a bit confused about the whole list starting at line 47. It's not completely clear to me how these items tie together. I.e., you talk about two material models, then in the next bullet you have a subbullet that specifies that you can use a range of material models. In the bullet "support for post-processing" it's not clear if items such as `ComputeResultantForceOnBoundary`, `ComputeMeanThermodynamicForcesValues`, etc are functions, classes, etc.
- line 63: you specify "the implementation of high ordermeshes or finite elements is straight-forward". Do you mean hat using higher order FEM with these nonlinear material models is made easy due to the use of the MFEM framework? I don't agree with the statement as is.
- line 63: nonlinear _solid_ mechanics? Although, you still have to apply boundary conditions for linear problems. Do you have to do something particularly different?
- line 65: What specifically do you mean by "advanced boundary conditions"?
- line 67: I'm not sure what you mean by a declarative "text based" API. Although all source code is text, we don't typically think of our APIs being text-based.
- line 102: you specify that you have a 30% overhead from MFEM/MGIS. Can you provide more details on what this test setup was, what hardware you were using, etc?
- line 113, MFEM has been scaled onto thousands of GPUs, for example see:  https://journals.sagepub.com/doi/10.1177/10943420231175462
