### 2025-02-28 thelfer/mfem-mgis#103 @jacobmerson
JOSS: functionality documentation
Do you have API level documenation?

https://github.com/openjournals/joss-reviews/issues/7719
