### 2025-02-28 thelfer/mfem-mgis#104 @jacobmerson
JOSS: Automated testing
Have you integrated an automated test suite? I was looking at your github actions and it seems like you just test the build with spack, but don't run tests. If so, please point me to the test suite.

Thanks!

https://github.com/openjournals/joss-reviews/issues/7719
