### 2025-02-28 #7719 @jacobmerson
@thelfer Thank you for your contribution to the community.

I have opened up some specific issues on the mfem-mgis github repo of open questions/items that need to be resolved.
