### 2025-02-28 thelfer/mfem-mgis#104 @rprat-pro
Spack tests are run regularly to check that the software stack is still sound. When we make a merge or commit, we have a complete compilation with cmake and run the tests available in the `tests` folder (using different parameter sets, 80 tests).

Tests: https://github.com/thelfer/mfem-mgis/actions/workflows/cmake.yml
