### 2025-02-28 thelfer/mfem-mgis#102 @rprat-pro
Hello @jacobmerson , thank you for all your comments about the paper.

About performance comments:

- [ ] "You mention that integration into MFEM will allow you to "efficiently use supercomputers". Have you done any scaling, or demonstration of this?" Tests are available into the documentation here over 16k cores: https://thelfer.github.io/mfem-mgis/benchmark.html .
- [ ] line 102: you specify that you have a 30% overhead from MFEM/MGIS. Can you provide more details on what this test setup was, what hardware you were using, etc? -> I'll add a simple and reproducible case available in MFEM and MFEM-MGIS.
- [ ] line 46, you need to demonstrate that you can "maintain parallel performance" -> I'm not sure I understand how to demonstrate this, most of the parallel computation time (>80% if not 95%) is taken up by solving linear systems when calling MFEM, so unless we make a big mistake in our implementation, performance should be maintained over time and benefit from future innovations in MFEM. Note that our developments with MGIS/MFront consist of PointWise operations which therefore have an algorithmic complexity in N (N being the number of integration points).
