+++ mfem-mgis/doxygen/dir_06124b4da483e359fc6eb505060e6b5b.js
@@ -1,25 +1,0 @@
-var dir_06124b4da483e359fc6eb505060e6b5b =
-[
-    [ "BuildFromRotationMatrix.hxx", "_build_from_rotation_matrix_8hxx.html", "_build_from_rotation_matrix_8hxx" ],
-    [ "ConvertKirchhoffStressDerivativeToSpatialModuli.hxx", "_convert_kirchhoff_stress_derivative_to_spatial_moduli_8hxx.html", null ],
-    [ "ConvertLogarithmicStrainTangentOperator.hxx", "_convert_logarithmic_strain_tangent_operator_8hxx.html", "_convert_logarithmic_strain_tangent_operator_8hxx" ],
-    [ "ConvertSpatialModuliToKirchhoffJaumanRateModuli.hxx", "_convert_spatial_moduli_to_kirchhoff_jauman_rate_moduli_8hxx.html", "_convert_spatial_moduli_to_kirchhoff_jauman_rate_moduli_8hxx" ],
-    [ "ConvertT2toST2ToST2toST2Expr.hxx", "_convert_t2to_s_t2_to_s_t2to_s_t2_expr_8hxx.html", "_convert_t2to_s_t2_to_s_t2to_s_t2_expr_8hxx" ],
-    [ "ConvertToTangentModuli.hxx", "_convert_to_tangent_moduli_8hxx.html", null ],
-    [ "st2tost2.ixx", "st2tost2_8ixx.html", "st2tost2_8ixx" ],
-    [ "ST2toST2Concept.hxx", "_s_t2to_s_t2_concept_8hxx.html", "_s_t2to_s_t2_concept_8hxx" ],
-    [ "ST2toST2Concept.ixx", "_s_t2to_s_t2_concept_8ixx.html", "_s_t2to_s_t2_concept_8ixx" ],
-    [ "ST2toST2ConceptIO.hxx", "_s_t2to_s_t2_concept_i_o_8hxx.html", null ],
-    [ "ST2toST2ConceptOperations.hxx", "_s_t2to_s_t2_concept_operations_8hxx.html", "_s_t2to_s_t2_concept_operations_8hxx" ],
-    [ "ST2toST2ConceptPushForward.ixx", "_s_t2to_s_t2_concept_push_forward_8ixx.html", "_s_t2to_s_t2_concept_push_forward_8ixx" ],
-    [ "st2tost2ResultType.hxx", "st2tost2_result_type_8hxx.html", "st2tost2_result_type_8hxx" ],
-    [ "ST2toST2ST2toST2ProductExpr.hxx", "_s_t2to_s_t2_s_t2to_s_t2_product_expr_8hxx.html", "_s_t2to_s_t2_s_t2to_s_t2_product_expr_8hxx" ],
-    [ "ST2toST2StensorProductExpr.hxx", "_s_t2to_s_t2_stensor_product_expr_8hxx.html", "_s_t2to_s_t2_stensor_product_expr_8hxx" ],
-    [ "ST2toST2TransposeExpr.hxx", "_s_t2to_s_t2_transpose_expr_8hxx.html", "_s_t2to_s_t2_transpose_expr_8hxx" ],
-    [ "StensorSquareDerivative.hxx", "_stensor_square_derivative_8hxx.html", "_stensor_square_derivative_8hxx" ],
-    [ "StensorST2toST2ProductExpr.hxx", "_stensor_s_t2to_s_t2_product_expr_8hxx.html", "_stensor_s_t2to_s_t2_product_expr_8hxx" ],
-    [ "StensorSymmetricProductDerivative.hxx", "_stensor_symmetric_product_derivative_8hxx.html", "_stensor_symmetric_product_derivative_8hxx" ],
[…]
