+++ mfem-mgis/doxygen/structmfront_1_1_material_property_query.js
@@ -1,9 +1,0 @@
-var structmfront_1_1_material_property_query =
-[
-    [ "MaterialPropertyQuery", "structmfront_1_1_material_property_query.html#a606a1e910829eb48bbe6ed0678b9ef1d", null ],
-    [ "~MaterialPropertyQuery", "structmfront_1_1_material_property_query.html#a390500a20b0406fad80ab7d512612c22", null ],
-    [ "exe", "structmfront_1_1_material_property_query.html#a898ab7b431376bc3087aa0168ec1bd14", null ],
-    [ "getDSL", "structmfront_1_1_material_property_query.html#a17638162bd51cc161de2d439b3491edb", null ],
-    [ "getNumberOfQueriesToBeTreated", "structmfront_1_1_material_property_query.html#a0ca998768a8d26a66ce00d7c11cf0119", null ],
-    [ "registerCommandLineCallBacks", "structmfront_1_1_material_property_query.html#a06c6882e075d37271d6bf0f3bd864b79", null ]
-];
