+++ mfem-mgis/doxygen/_abstract_test_case_factory_8hxx.js
@@ -1,4 +1,0 @@
-var _abstract_test_case_factory_8hxx =
-[
-    [ "mfmtg::AbstractTestCaseFactory", "structmfmtg_1_1_abstract_test_case_factory.html", "structmfmtg_1_1_abstract_test_case_factory" ]
-];
