+++ mfem-mgis/doxygen/dir_1e3d02143978375a506ff49b4fb58c3b.js
@@ -1,19 +1,0 @@
-var dir_1e3d02143978375a506ff49b4fb58c3b =
-[
-    [ "AnalyseJunction.hxx", "_analyse_junction_8hxx.html", null ],
-    [ "BCC.hxx", "_b_c_c_8hxx.html", "_b_c_c_8hxx" ],
-    [ "Config.hxx", "spack_2opt_2spack_2linux-ubuntu24_804-zen2_2gcc-12_83_80_2tfel-master-udmvhtlh7nuq6vwaqycq7tcpiqb0c58dcaad355ca37d74c88a1104ad2b.html", null ],
-    [ "Crystallo.hxx", "_crystallo_8hxx.html", "_crystallo_8hxx" ],
-    [ "Cubic.hxx", "_cubic_8hxx.html", "_cubic_8hxx" ],
-    [ "FCC.hxx", "_f_c_c_8hxx.html", "_f_c_c_8hxx" ],
-    [ "FrankRead.hxx", "_frank_read_8hxx.html", "_frank_read_8hxx" ],
-    [ "GSystem.hxx", "_g_system_8hxx.html", "_g_system_8hxx" ],
-    [ "Hardening.hxx", "_hardening_8hxx.html", "_hardening_8hxx" ],
-    [ "HCP.hxx", "_h_c_p_8hxx.html", "_h_c_p_8hxx" ],
-    [ "IBurgers.hxx", "_i_burgers_8hxx.html", "_i_burgers_8hxx" ],
-    [ "IDirection.hxx", "_i_direction_8hxx.html", "_i_direction_8hxx" ],
-    [ "IPlane.hxx", "_i_plane_8hxx.html", "_i_plane_8hxx" ],
-    [ "IsotropicLineTensionModel.hxx", "_isotropic_line_tension_model_8hxx.html", "_isotropic_line_tension_model_8hxx" ],
-    [ "TripleNode.hxx", "_triple_node_8hxx.html", "_triple_node_8hxx" ],
-    [ "Vect3.hxx", "_vect3_8hxx.html", "_vect3_8hxx" ]
-];
