+++ mfem-mgis/doxygen/_jacobian_type_8hxx.js
@@ -1,7 +1,0 @@
-var _jacobian_type_8hxx =
-[
-    [ "tfel::config::internals::JacobianType< N, T, use_qt >", "structtfel_1_1config_1_1internals_1_1_jacobian_type.html", null ],
-    [ "tfel::config::internals::JacobianType< N, T, false >", "structtfel_1_1config_1_1internals_1_1_jacobian_type_3_01_n_00_01_t_00_01false_01_4.html", null ],
-    [ "tfel::config::internals::JacobianType< 1u, T, use_qt >", "structtfel_1_1config_1_1internals_1_1_jacobian_type_3_011u_00_01_t_00_01use__qt_01_4.html", null ],
-    [ "tfel::config::internals::JacobianType< 1u, T, false >", "structtfel_1_1config_1_1internals_1_1_jacobian_type_3_011u_00_01_t_00_01false_01_4.html", null ]
-];
