+++ mfem-mgis/doxygen/dir_46fc3e526c9ba2e1ac20739f1d833d7a.js
@@ -1,5 +1,0 @@
-var dir_46fc3e526c9ba2e1ac20739f1d833d7a =
-[
-    [ "Expr.hxx", "_expression_templates_2_expr_8hxx.html", "_expression_templates_2_expr_8hxx" ],
-    [ "StandardOperations.hxx", "_standard_operations_8hxx.html", "_standard_operations_8hxx" ]
-];

[…]
