+++ mfem-mgis/doxygen/_johnson_cook_strain_rate_sensitivity_factor_8hxx.js
@@ -1,0 +1,4 @@
+var _johnson_cook_strain_rate_sensitivity_factor_8hxx =
+[
+    [ "mfront::bbrick::JohnsonCookStrainRateSensitivityFactor", "structmfront_1_1bbrick_1_1_johnson_cook_strain_rate_sensitivity_factor.html", "structmfront_1_1bbrick_1_1_johnson_cook_strain_rate_sensitivity_factor" ]
+];
