+++ mfem-mgis/doxygen/structtfel_1_1math_1_1_derivative_type_dispatcher_3_01_vector_tag_00_01_vector_tag_00_01tvector_c6da30a2eb349e2b8f7044ab4352d283.js
@@ -1,0 +1,4 @@
+var structtfel_1_1math_1_1_derivative_type_dispatcher_3_01_vector_tag_00_01_vector_tag_00_01tvector_c6da30a2eb349e2b8f7044ab4352d283 =
+[
+    [ "type", "structtfel_1_1math_1_1_derivative_type_dispatcher_3_01_vector_tag_00_01_vector_tag_00_01tvector_c6da30a2eb349e2b8f7044ab4352d283.html#a17c223f6f8f0ffa4dc82c209fb89aa33", null ]
+];
