+++ mfem-mgis/doxygen/structnumodis_1_1_i_plane.js
@@ -1,0 +1,12 @@
+var structnumodis_1_1_i_plane =
+[
+    [ "IPlane", "structnumodis_1_1_i_plane.html#ab3f37c4a5f0a2b752108ef3ba9c3bdb9", null ],
+    [ "IPlane", "structnumodis_1_1_i_plane.html#a5f6a53fa5d411aa48bd6b0235e82d4b5", null ],
+    [ "getIndex", "structnumodis_1_1_i_plane.html#a44c115863d9e6d862a2dae04270ffb2a", null ],
+    [ "getIndex", "structnumodis_1_1_i_plane.html#a954886a8bf2b6ca4fb563c0007b2cec8", null ],
+    [ "getNindices", "structnumodis_1_1_i_plane.html#acbde8a378f89ed731a867179f976efe6", null ],
+    [ "IsNull", "structnumodis_1_1_i_plane.html#ad8cef38e63baa01bda90accf7dc9db91", null ],
+    [ "operator[]", "structnumodis_1_1_i_plane.html#af35b969209e2851c97bf6fe8e00a3066", null ],
+    [ "setIPlane", "structnumodis_1_1_i_plane.html#a760b7afcb2d945548b6ae10801a34d17", null ],
+    [ "_index", "structnumodis_1_1_i_plane.html#a1205088146ac3ad496a5fe612a0ec674", null ]
+];
