### 2025-03-03 thelfer/mfem-mgis#104 @jacobmerson
Thanks. I will do that locally. It would be good if you could also add this to your github actions configuration.
