+++ .github/workflows/cmake.yml
@@ -7,7 +7,7 @@ on:
     branches: [ "master" ]
 
 jobs:
-  build:
+  build-and-run:
 
     runs-on: ubuntu-latest
 
