+++ .github/workflows/cmake.yml
@@ -1,4 +1,4 @@
-name: Cmake
+name: Build with Cmake and Run Examples
 
 on:
   push:
@@ -7,7 +7,7 @@ on:
     branches: [ "master" ]
 
 jobs:
-  build_and_run_tests:
+  run_tests:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
