### 2025-03-03 thelfer/mfem-mgis#104 @rprat-pro
The job action was named `build` (for `cmake.yaml`) but there was a step called `Run examples` which executes this command. For the sake of clarity, I have renamed this workflow to `Build with Cmake and Run Examples` and the job name to `run_tests`.
