### 2025-03-04 cb3be37
refactor paper
