+++ docs/paper/joss/paper.md
@@ -32,7 +32,9 @@ bibliography: paper.bib
 
 The `MFEM/MGIS` application aims at efficiently use supercomputers in
 order to describe coupled multiphysics phenomena with a particular focus
-on thermo-mechanics.
+on thermo-mechanics. The authors primarily aims at describing the
+nuclear fuels at mesoscale (see example below), but `MFEM/MGIS` is
+versatile and can address more general cases.
 
 This open-source library is based on several components as
 prerequisites: the `MFEM` (Modular Finite Element Methods) [@mfem;
@@ -40,14 +42,11 @@ prerequisites: the `MFEM` (Modular Finite Element Methods) [@mfem;
 [@Helfer2020] library and the `MFront` code generator
 [@helfer2015introducing].
 
-Thanks to the features embedded within MGIS and MFront and thanks to
-specific developments, MFEM/MGIS adds several mechanical features
+Thanks to the features embedded within `MGIS` and `MFront` and thanks to
+specific developments, `MFEM/MGIS` adds several mechanical features
 compared to a pure MFEM approach. The library tackles some peculiarities
 of nonlinear mechanics. In particular, the support of complex
 constitutive laws and the management of advanced boundary conditions.
-
-It provides a high level of abstraction based on a declarative
-text-based Application Programming Interface suitable for engineers.
 
 # MFEM-based Thermo-Mechanical Solver for Nuclear Fuel Simulations
 
@@ -80,66 +79,81 @@ library handles all the kinds of behavior supported by the open-source
 In the field of nonlinear mechanics, this encompasses arbitrary complex
 behaviors that can describe damage, plasticity, viscoplasticity, phase
 change in both small or finite strain analyses. Generalized behaviors
-such as variational approaches to fracture, Cosserat media are supported
-by `MFEM/MGIS`. Strongly coupled thermo-mechanical behaviors and
-Cosserat media can be introduced by plugins and will be integrated in
-future versions of `MFEM/MGIS`.
+such as variational approaches to fracture are supported
+by `MFEM/MGIS`.
 
-Through its integration with `MGIS` and `MFront`, along with dedicated
-advancements, `MFEM/MGIS` has incorporated the subsequent mechanical
-attributes in contrast to the pure `MFEM` approach while maintaining
-parallel performance:
+Also, contrary to non linear forms provided natively by `MFEM`,
+`MFEM/MGIS` allows to assign distinct behaviors to each material. The
+`MGIS` data structures are also use to add support for partial
+quadrature functions to `MFEM`, a feature needed to store internal state
+variables on each material.
 
-- Ability to handle several materials which distinct constitutive equations:
-  - Micromorphique laws
-  - Orthotropic laws 
-- Support for internal state variables (defined at integration points):
-  - Use of `MGIS` data structures
-  - Use of mechanical behavior laws generated by `MFront`
-- Support for post-processing specific to nonlinear mechanic:
-  - `ComputeResultantForceOnBoundary`: Compute the resultant of the inner
-    forces on a boundary
-  - `ComputeMeanThermodynamicForcesValues`: Compute the macroscopic stress
-    and strain for each materials
-  - `ParaviewExportIntegrationPointResultsAtNodes`: Paraview post
-    processing files with internal state variables according to `MGIS`
-- Flexible C++ API to easily define MFEM simulation features
+<!--
+Strongly coupled thermo-mechanical behaviors and Cosserat media can be
+introduced by plugins and will be integrated in future versions of
+`MFEM/MGIS`.
+-->
+
+# Overview of `MFEM/MGIS` features
+
+## User interface
+
+The `MFEM/MGIS` library is written in `C++17` language.
+
[…]
