+++ docs/paper/joss/paper.md
@@ -24,7 +24,7 @@ authors:
 affiliations:
  - name: CEA, DES, IRESNE, DEC, Cadarache F 13108 St Paul Lez Durance
    index: 1
-date: 3 August 2023
+date: 6 March 2025
 bibliography: paper.bib
 ---
 
@@ -48,15 +48,16 @@ compared to a pure MFEM approach. The library tackles some peculiarities
 of nonlinear mechanics. In particular, the support of complex
 constitutive laws and the management of advanced boundary conditions.
 
-# MFEM-based Thermo-Mechanical Solver for Nuclear Fuel Simulations
+# About MFEM
 
-`MFEM`, is a finite element library designed for current supercomputers
-but also for the upcoming exascale supercomputers. It provides many
-useful features for carrying out realistic simulations: support for
-curvilinear meshes, high order approximation spaces and different
-families of finite elements, interfaces to several types of parallel
-solvers (including matrix-free ones), preconditioners, and native
-support for adaptive non-conforming mesh refinement (AMR).
+`MFEM`, is a finite element library designed for current
+supercomputers but also for the upcoming exascale supercomputers. It
+provides many useful features for carrying out state-of-the art
+simulations: support for curvilinear meshes, high order approximation
+spaces and different families of finite elements, interfaces to
+several types of parallel solvers (including matrix-free ones),
+preconditioners, and native support for adaptive non-conforming mesh
+refinement (AMR).
 
 Originating from the applied mathematics and parallel computing
 communities, `MFEM` offers both performance and a large panel of
@@ -67,9 +68,11 @@ microstructure and mesoscale modeling for nuclear fuel.
 
 # Statement of need
 
-The solid mechanic applications in `MFEM` are mostly limited to simple
-constitutive equations such as elasticity and hyperelasticity, which is
-insufficient to address complex nuclear fuel simulations.
+The solid mechanic examples in `MFEM` are mostly limited to simple
+constitutive equations such as elasticity and hyperelasticity without
+internal state variables. This is insufficient to address many
+engineering studies and in particular complex nuclear fuel
+simulations.
 
 The aim of `MFEM/MGIS` project is to combine `MFEM` with the
 `MFrontGenericInterfaceSupport` (`MGIS`) project, an open-source `C++`
@@ -77,14 +80,11 @@ library handles all the kinds of behavior supported by the open-source
 `MFront` code generator.
 
 In the field of nonlinear mechanics, this encompasses arbitrary complex
-behaviors that can describe damage, plasticity, viscoplasticity, phase
-change in both small or finite strain analyses. Generalized behaviors
-such as variational approaches to fracture are supported
-by `MFEM/MGIS`.
+behaviors that can describe damage, plasticity, viscoplasticity in both
+small or finite strain analyses. Generalized behaviors such as
+variational approaches to fracture are supported by `MFEM/MGIS`.
 
-Also, contrary to non linear forms provided natively by `MFEM`,
-`MFEM/MGIS` allows to assign distinct behaviors to each material. The
-`MGIS` data structures are also used to add support for partial
+The `MGIS` data structures are also used to add support for partial
 quadrature functions to `MFEM`, a feature needed to store internal state
 variables on each material.
 
@@ -96,6 +96,108 @@ introduced by plugins and will be integrated in future versions of
 
 # Overview of `MFEM/MGIS` features
 
+## The `NonLinearEvolutionProblem` class
+
+The main class of `MFEM/MGIS` is called `NonLinearEvolutionProblem` and
+describes the evolution of the materials of the physical system
+of interest over a single time step for a given phenomenon.
+
+Currently `MFEM/MGIS` provides built-in support for mechanics, heat
[…]
