### 2025-03-06 f681dc9
add 'State of the field' section
