### 2025-03-06 a6fa6d4
add mfem example
