+++ CMakeLists.txt
@@ -133,3 +133,4 @@ mfem_mgis_buildenv()
 add_subdirectory(include)
 add_subdirectory(src)
 add_subdirectory(tests)
+add_subdirectory(benchmarks)
