+++ benchmarks/CMakeLists.txt
@@ -6,10 +6,9 @@ if(MFEM_MGIS_HAVE_TFEL)
 			${PROJECT_SOURCE_DIR}/benchmarks/beam-tet.mesh
 			DESTINATION ${PROJECT_BINARY_DIR})
 
+# MFEM-MGIS example
 	add_executable(MFEMMGISLinearElasticityBenchmark
 			MFEMMGISLinearElasticityBenchmark.cxx)
-
-
 	add_custom_command(
 			OUTPUT "src/libBehaviour.so"
 			COMMAND mfront --obuild --interface=generic ${PROJECT_BINARY_DIR}/Elasticity.mfront
@@ -24,5 +23,14 @@ if(MFEM_MGIS_HAVE_TFEL)
 			DEPENDS "src/libBehaviour.so")
 
   add_dependencies(MFEMMGISLinearElasticityBenchmark generate-libBehaviour-benchmark)
+
+# MFEM example
+	add_executable(MFEMLinearElasticityBenchmark
+			MFEMLinearElasticityBenchmark.cxx)
+#  target_include_directories(MFEMLinearElasticityBenchmark
+#    PUBLIC ${MFEM_INCLUDE_DIRS})
+
+  target_link_libraries(MFEMLinearElasticityBenchmark
+    PUBLIC MFEMMGIS)
 
 endif()
