### 2025-03-06 db4e25e
fix fig for the joss paper
