+++ docs/paper/joss/paper.bib
@@ -182,7 +182,7 @@ title = {Parmetis: Parallel graph partitioning and sparse matrix ordering librar
 	pages = {103874},
 	year = {2020},
 	issn = {0022-5096},
-	doi = {https://doi.org/10.1016/j.jmps.2020.103874},
+	doi = {10.1016/j.jmps.2020.103874},
 	url = {https://www.sciencedirect.com/science/article/pii/S0022509619306453},
 	author = {Renaud Masson and Mohamed El Bachir Seck and Jules Fauque and Mihail Gărăjeu},
 	keywords = {Nonlinear viscoelasticity, Elasto-viscoplasticity, Two-phase composites, Hashin–Shtrikman, Stress-free strain, Second-order moments, FFT calculations},
