### 2025-03-06 0707834
add an option in the CMakelist to avoid the compilation of the benchmarks
