### 2025-03-07 4753a45
Update benchmark.rst
