### 2025-03-07 thelfer/mfem-mgis#102 @thelfer
First of all, I must thank you for your remarks. They led us to refactor
the paper and rewrite it to a great extent.

You'll find attach the new version of the paper which is not yet merged
in `master` nor in the branch with the official submission. *We would be
thankful if we could have your impressions before doing so*.

[paper.pdf](https://github.com/user-attachments/files/19124292/paper.pdf)

> The paper focused on the integration of MGIS with MFEM and had
> essentially no reference to nuclear fuels or mesoscale simulations. This
> isn't necessarily a problem as I think the integration with MGIS is a
> useful contribution on its own. But, either the title should remove the
> discussion of nuclear fuels, or the text should provide more relevant
> background.

This is a very interesting remark.

The main driving force for the development of `MFEM/MGIS` is the need,
for the research made in our department, to make nonlinear
thermo-mechanical simulations of nuclear fuels at mesoscale.

However, `MFEM/MGIS` is versatile and could be used for more general
thermo-mechanical computations.

The fact is that the paper evolved a lot. In an early version, we did
describe two tests cases, but due to constraints on the length of the
paper, we removed them. We added a picture of a simulation of a RVE of
MOx that reproduces results obtained with FFT in a « recent » paper.

> There are many finite element solvers for nonlinear solid mechanics
> problems. The paper is missing a discussion of how this relates to any
> of those. E.g., Moose, code-aster, abaqus, ansys, etc. Some additional
> discussion on what is MFront and how that generalizes the integration
> of material models into MFEM is needed.

We added a "State of the field" section to explain why we choose to
develop `MFEM/MGIS`.

Concerning `MFront`, its role is now more explicit throughout the paper.
Let us know if it seems satisfying.

> You mention that integration into MFEM will allow you to "efficiently
> use supercomputers". Have you done any scaling, or demonstration of
>  this?

We strengthened the « Numerial results » section. We also added a link
to some benchmarks available in the documentation:
https://thelfer.github.io/mfem-mgis/benchmark.html.

> The way the first sentence in the statement of need is written (line
>  33) is misleading. MFEM is a finite element framework and doesn't make
>  any assumptions about the types of constitutive laws that you
>  implement.

`MFEM` does not provide, out of the box, any example of mechanical
behaviour with internal state variables: the examples provided are
restricted to linear elasticity and simple hyperelastic behaviours with
uniform properties per material. We added a note about how our solution
differs.

We also refactored the paper to give more insights to the extensions
that we made to `MFEM`, emphasising on :

- partial quadrature functions,
- flexibility brought by behaviour integrators,
- Etc..

> line 48 micromorphique -> micromorphic

Fixed

> You specifically mention micromorphic and orthotropic material laws.
> Is there a limitation in the current implementation that prevents
> other laws that are implemented in MFront? You also mention plasticity
> in the introduction, is this included?

There is no limitation in the kind of mechanical behaviours that can be
used in `MFEM/MGIS`. In mechanical problems, you can use any finite
strain and small strain behaviours that can be implemented in `MFront`.
So there is in practice no restriction on the complexity of the
behaviour that can be used in `MFEM/MGIS`.

One key feature is that MFEM/MGIS does not define any behaviour : all
beahviours are loaded at runtime and they must have been compiled thanks
to `MFront` before running a simulation with `MFEM/MGIS`. This is now
illustrated by one of the code snipet that we added.

> I'm a bit confused about the whole list starting at line 47. It's not
> completely clear to me how these items tie together. I.e., you talk
> about two material models, then in the next bullet you have a
> subbullet that specifies that you can use a range of material models.
> In the bullet "support for post-processing" it's not clear if items
> such as ComputeResultantForceOnBoundary,
> ComputeMeanThermodynamicForcesValues, etc are functions, classes, etc.

This list was indeed badly written and unclear. We refactored the paper
to describe in dedicated paragraphs the key features introduced in
`MFEM/MGIS`.

> you specify "the implementation of high ordermeshes or finite
> elements is straight-forward". Do you mean hat using higher order FEM
> with these nonlinear material models is made easy due to the use of
> the MFEM framework? I don't agree with the statement as is.

Thanks to MFEM, we can change the order of the finite element space :
this does work out of the box. We now mention this in the paper.

> nonlinear solid mechanics? Although, you still have to apply
> boundary conditions for linear problems. Do you have to do something
> particularly different?

I am not sure to understand this question. It is worth mentioning that
most of the boundary conditions of interest for our use cases (see next
question) can’t be used in linear problems.

> What specifically do you mean by "advanced boundary conditions"?

We mean contact, multiple-points contraints, pressure (in finite strain
analyses) for instance. Those boundary conditions are yet to be
implemented in `MFEM/MGIS`, but we already set up the interface to
introduce those boundary conditions.

> I'm not sure what you mean by a declarative "text based"
>  API. Although all source code is text, we don't typically think of our
>  APIs being text-based.

This sentence was elusive. We tried to emphasize the fact that most
objects are instantiated thanks to a specific data structure which is a
C++ equivalent to a python dictionary. Internally this structure is
passed to abstract factories.We already used this approach in other
projects and this allowed to easily create input files and create python
bindings.All in all, this approach leads to a code with is close to
so-called « literate programming »

> you specify that you have a 30% overhead from MFEM/MGIS.
> Can you provide more details on what this test setup was, what
> hardware you were using, etc?

This statement was dubious. We tried to make a clearer statement in a
dedicated section (« about linear resolutions »). In addition, a
benchmark has been added into the benchmarks folder containing a
comparison between MFEM and MFEM-MGIS. In this example we highlight
around of 30% of overhead with MFEM-MGIS on our supercomputer but we
can’t affirm that it’s always true for every problems and meshes, so we
finally decided to remove this statement.

> MFEM has been scaled onto thousands of GPUs, for example
> see: https://journals.sagepub.com/doi/10.1177/10943420231175462

We removed the conclusion and anything with is more a perspective than a
proven (and not potential) feature of `MFEM/MGIS`.
