### 2025-03-09 thelfer/mfem-mgis#102 @jacobmerson
I think this is much better and really gets at the core of the power of your approach. There are a few typos/half completed sentences that need to get cleaned up.

I do have one remaining technical question:

Can you explain what you mean by this?
> Contrary to non linear forms provided natively by MFEM, MFEM/MGIS allows to assign distinct
behaviors to each material.

Depending on what you wanted to do, you can apply different behaviors or different properties in the mfem elasticity integration. For example, if you just want to change the elasticity properties in different parts of your model you can pass in a `PiecewiseConstCoefficient`, or `FunctionCoefficient`

If you wanted to apply different material models in different elements, you can use the version of `AddDomainIntegrator` that lets you apply the bilinear form to a subset of the finite element mesh.
