### 2025-03-09 thelfer/mfem-mgis#102 @thelfer
@jacobmerson Thanks for this nice feedback.

> There are a few typos/half completed sentences that need to get cleaned up.

Yes, we need to refine this new version, but we wanted to have your feedbacks .

> Depending on what you wanted to do, you can apply different behaviors or different properties in the mfem elasticity integration.

I think that this statement covers this:

> For instance, MFEM elastic kernels assume that:
>
> - all materials are elastic,

In MFEM/MGIS, some material can be elastic, other can be plastic, etc...

> If you wanted to apply different material models in different elements, you can use the version of AddDomainIntegrator that lets you apply the bilinear form to a subset of the finite element mesh.

I think that this overload was not available when we started MFEM/MGIS. I will dig further to see if this is also supported in nonlinear resolutions. Let me check
