### 2025-03-09 thelfer/mfem-mgis#102 @thelfer
> I think that this overload was not available when we started MFEM/MGIS. I will dig further to see if this is also supported in nonlinear resolutions. Let me check

Ok, this overload is also available in nonlinear forms. I'll change the text accordingly.
