+++ docs/paper/joss/paper.md
@@ -161,23 +161,17 @@ linear behaviours (generated by `MFront`), but with a computational
 overhead compared to linear analysis made with optimised kernels, such
 as the elastic kernels provided natively by `MFEM`.
 
-In our experience, this overhead is limited and mostly comes from
-the extra flexibility allowed by `MFEM/MGIS`. For instance,
-`MFEM` elastic kernels assume that:
-
-- all materials are elastic,
-- material properties (Young's modulus, Poisson's ratio) are uniform
-  on each material.
-
-`MFEM/MGIS` kernels don't have those restrictions as described below.
+In our experience, this overhead is limited and mostly comes from the
+extra flexibility allowed by `MFEM/MGIS`. For instance, `MFEM` elastic
+kernels assume that material properties (Young's modulus, Poisson's
+ratio) are uniform on each material.
 
 ## Behaviour integrators
 
-Contrary to non linear forms provided natively by `MFEM`, `MFEM/MGIS`
-allows to assign distinct behaviors to each material. To achieve this, a
-special nonlinear formulation has been implemented which delegates the
-computations of residual and jacobian terms on each material to
-so-called behaviour integrators.
+`MFEM/MGIS` allows to assign distinct behaviors to each material. To
+achieve this, a special nonlinear formulation has been implemented which
+delegates the computations of residual and jacobian terms on each
+material to so-called behaviour integrators.
 
 Behaviour integrators are associated with a physical phenomenon and a
 modelling hypothesis (plane strain, plane stress).
@@ -252,7 +246,6 @@ Several examples can be found on the open-source GitHub repository:
 <https://github.com/latug0/mfem-mgis-examples>, including simulation of
 the microstructural scale using a Representative Volume Element (RVE) of
 nuclear fuel, figure \autoref{fig:RVE}. 
-
 
 ![Simulation of a Representative Volume Element (RVE) of mixed oxide material, which contains 634 inclusions. The variable "u" represents the displacement. This simulation reproduces the results obtained by [@MASSON2020103874], who used an FFT method. \label{fig:RVE}](./MOX17percent634Spheres.png){width=70%}
 
