### 2025-03-09 af2fc62
exclude forward declarations of MFEM structures
