+++ mfem-mgis/doxygen/wstream_view_8hxx.js
@@ -1,0 +1,5 @@
+var wstream_view_8hxx =
+[
+    [ "tfel::system::stream_traits< wstreamView< is > >", "structtfel_1_1system_1_1stream__traits_3_01wstream_view_3_01is_01_4_01_4.html", null ],
+    [ "tfel::system::wstreamView< isBlocking >", "structtfel_1_1system_1_1wstream_view.html", null ]
+];
