+++ .github/workflows/doxygen.yml
@@ -66,4 +66,5 @@ jobs:
         personal_token: ${{ secrets.GITHUB_TOKEN }}
         publish_branch: master
         publish_dir: ${{github.workspace}}/build/docs/doxygen/html/ 
+        external_repository: thelfer/thelfer.github.io
         destination_dir: mfem-mgis/doxygen/
