+++ mfem-mgis/doxygen/_real_part_type_8hxx.js
@@ -1,6 +1,0 @@
-var _real_part_type_8hxx =
-[
-    [ "tfel::typetraits::RealPartType< T >", "structtfel_1_1typetraits_1_1_real_part_type.html", null ],
-    [ "TFEL_TYPETRAITS_REALPARTTYPE", "_real_part_type_8hxx.html#a56539a6309aa618368fe2f278fb31a0a", null ],
-    [ "TFEL_TYPETRAITS_REALPARTTYPEII", "_real_part_type_8hxx.html#aa85acf2a9798c7a9896566b73d2e019f", null ]
-];
