+++ mfem-mgis/doxygen/_europlexus_stress_free_expansion_handler_8hxx.js
@@ -1,5 +1,0 @@
-var _europlexus_stress_free_expansion_handler_8hxx =
-[
-    [ "EuroplexusLogarithmicStrainStressFreeExpansionHandler", "_europlexus_stress_free_expansion_handler_8hxx.html#a34414c3c61ffee64e894781deedf762b", null ],
-    [ "EuroplexusStandardSmallStrainStressFreeExpansionHandler", "_europlexus_stress_free_expansion_handler_8hxx.html#a683f1ed106fb1376bd93f052bff8b0c7", null ]
-];
