+++ mfem-mgis/doxygen/search/variables_19.js
@@ -1,6 +1,0 @@
-var searchData=
-[
-  ['y_0',['y',['../structtfel_1_1math_1_1_cubic_spline_collocation_point.html#ac203a752fa6287aea8052e50acadeb0a',1,'tfel::math::CubicSplineCollocationPoint::y'],['../structtfel_1_1math_1_1_tiny_runge_kutta_solver_base.html#ab6129fa5e05b733bb4c3e899199e2757',1,'tfel::math::TinyRungeKuttaSolverBase::y']]],
-  ['yellow_1',['Yellow',['../structtfel_1_1utilities_1_1_terminal_colors.html#ac1ec81a815a8dcc4c109b11521c6f677',1,'tfel::utilities::TerminalColors']]],
-  ['young_5fmodulus_2',['young_modulus',['../structmtest_1_1_mistral_behaviour.html#a6a91b752f0c933771d1e779738e058b5',1,'mtest::MistralBehaviour']]]
-];

[…]
