+++ docs/paper/joss/paper.md
@@ -64,7 +64,8 @@ refinement (AMR).
 Originating from the applied mathematics and parallel computing
 communities, `MFEM` offers both performance and a large panel of
 advanced mathematical features. In particular, one can easily switch
-from one linear solver to another (direct or iterative), essential for the targeted application [@bernaud2024pleiades]:
+from one linear solver to another (direct or iterative),
+which is essential for the targeted application [@bernaud2024pleiades]:
 microstructure and mesoscale modeling for nuclear fuel.
 
 ## Statement of need
@@ -75,14 +76,14 @@ internal state variables. This is insufficient to address many
 engineering studies and in particular complex nuclear fuel
 simulations.
 
-The aim of `MFEM/MGIS` project is to combine `MFEM` with the
+The aim of the `MFEM/MGIS` project is to combine `MFEM` with the
 `MFrontGenericInterfaceSupport` (`MGIS`) project, an open-source `C++`
-library handles all the kinds of behavior supported by the open-source
+library that handles all the kinds of behaviours supported by the open-source
 `MFront` code generator.
 
 In the field of nonlinear mechanics, this encompasses arbitrary complex
-behaviors that can describe damage, plasticity, and viscoplasticity in both
-small or finite strain analyses. Generalized behaviors such as
+behaviours that can describe damage, plasticity, and viscoplasticity in both
+small or finite strain analyses. Generalized behaviours such as
 variational approaches to fracture are supported by `MFEM/MGIS`.
 
 The `MGIS` data structures are used to add support for partial
@@ -91,7 +92,7 @@ variables on each material.
 
 ## State of the field
 
-Many open-source thermomechanical solvers allow to handle complex
+Many open-source thermomechanical solvers allow handling complex
 mechanical behaviours. `code_aster`, `MoFEM`, `CalculiX` are examples
 of state of the art solvers which have an interface with `MFront`.
 
@@ -103,7 +104,7 @@ performance computing and adapt it to engineering needs and evaluate
 the resulting performances.
 
 <!--
-Strongly coupled thermo-mechanical behaviors and Cosserat media can be
+Strongly coupled thermo-mechanical behaviours and Cosserat media can be
 introduced by plugins and will be integrated in future versions of
 `MFEM/MGIS`.
 -->
@@ -117,7 +118,7 @@ describes the evolution of the materials of the physical system
 of interest over a single time step for a given phenomenon.
 
 Currently `MFEM/MGIS` provides built-in support for mechanics, heat
-transfer, and micromorphic damage are supported.
+transfer, and micromorphic damage.
 
 The following snippet declares a new nonlinear evolution problem:
 
@@ -135,7 +136,7 @@ As the unknown is scalar (according to the `UnknownsSize` parameter),
 this problem can be used to describe heat transfer or micromorphic
 damage, depending on the behaviour integrators declared, as explained in
 the next section. The `NonLinearEvolutionProblem` class supports both
-sequential and parallel computations and let the user exploits a large
+sequential and parallel computations and lets the user exploit a large
 subset of `MFEM` abilities, including the use of finite elements of
 arbitrary orders.
 
@@ -148,7 +149,7 @@ several instances of `NonLinearEvolutionProblem`.
 `NonLinearEvolutionProblem` for periodic computations named
 `PeriodicNonLinearEvolutionProblem`.
 
-This class allows to manage the evolutions of the macroscopic
+This class allows managing the evolutions of the macroscopic
 gradients (strain in small strain analysis, deformation gradient in
 finite strain analysis, temperature gradient in heat transfer
 analysis) and pass them to the behaviour integrators.
[…]
