+++ .github/workflows/doxygen.yml
@@ -66,5 +66,5 @@ jobs:
         personal_token: ${{ secrets.DEPLOY_TO_THELFER_GITHUB_IO }}
         publish_branch: master
         publish_dir: ${{github.workspace}}/build/docs/doxygen/html/ 
-        external_repository: thelfer/thelfer.github.io
+#        external_repository: thelfer/thelfer.github.io
         destination_dir: mfem-mgis/doxygen/
