### 2025-03-14 #7719 @thelfer
@arfon For your information, thanks to @jacobmerson's review, the paper has been greatly improved.

@iammix, you may want to review the changes: here https://github.com/thelfer/mfem-mgis/issues/102
