+++ docs/source/index.rst
@@ -1,30 +1,46 @@
+.. _mfem_mgis:
+
 =========================
 The ``MFEM/MGIS`` project
 =========================
 
+.. toctree::
+   :hidden:
+   :maxdepth: 1
+
+   user_guide/user_guide.rst
+   commented_examples/commented_examples.rst
+   developer_guide/developer_guide.rst
+   installation_guide/installation_guide.rst
+
 The aim of the ``MFEM/MGIS`` project is to provide a ``C++`` library to
 build advanced mechanical simulation based on ``MFront`` behaviours with
-``MFEM`` :cite:`Anderson_2020`. The ``MGIS`` project is used to interface
-``MFront`` behaviours :cite:`helfer_mfrontgenericinterfacesupport_2020`.
+``MFEM`` . The ``MGIS`` project is used to interface
+``MFront`` behaviours .
 
-This project is written in ``C++-17``.
+The ``MFEM-MGIS`` project, aims at efficiently use supercomputers in
+order to describe coupled multiphysics phenomena with a particular focus
+on thermo-mechanics. This open-source library is based on several
+components as prerequisites:
 
-.. _mfemmgis:
+- the ``mfem`` (Modular Finite Element Methods) library
+  :cite:`Anderson_2020`,
+- the ``MGIS`` (MFront Generic Interface Support) library
+  :cite:`helfer_mfrontgenericinterfacesupport_2020`,
+-  the ``MFront`` code generator.
 
-.. toctree::
-   :name: MFEM-MGIS Project
-   :numbered:
-   :maxdepth: 2      
-   :caption: MFEM MGIS Project
+Thanks to the features embedded within ``mgis`` and ``MFront`` and
+thanks to specific developments, ``MFEM-MGIS`` adds several mechanical
+features compared to a pure ``mfem`` approach.
 
-   about.rst
-   install.rst
-   changelog.rst
-   glossary.rst
-   post-processing.rst
-   tutorial.rst
-   examples.rst
-   benchmark.rst
+The library tackles some peculiarities of nonlinear mechanics. In
+particular, the support of complex constitutive laws and the management
+of advanced boundary conditions. It provides a high level of abstraction
+based focused on the physics to be treated.
 
+.. figure:: img/mfem-mgis-illustration.png
+
+References
+==========
 
 .. bibliography:: bibliography.bib
