+++ docs/source/developer_guide/developer_guide.rst
@@ -1,0 +1,13 @@
+.. _mfem_mgis_developer_guide:
+
+===============
+Developer guide
+===============
+
+This project is written in ``C++-17``.
+
+.. toctree::
+
+   benchmark.rst
+   changelog.rst
+   contributors.rst
