+++ docs/source/user_guide/post-processing.rst
@@ -2,7 +2,12 @@
 Post Processing Features
 ========================
 
-In this section, we describe the post-processing options available in MFEM-MGIS.
+In this section, we describe the post-processing options available in
+MFEM-MGIS.
+
+.. contents::
+    :depth: 3
+    :local:
 
 .. warning::
 
@@ -22,7 +27,7 @@ Dump Paraview Results
 
 **Results**
 
-.. figure:: _static/SatohTest.png
+.. figure:: img/SatohTest.png
 
 Export Integration Point Results At Nodes
 ------------------------------------------
@@ -43,11 +48,11 @@ Export Integration Point Results At Nodes
 
 **Results**
 
-.. figure:: _static/SatohTestStress.png
+.. figure:: img/SatohTestStress.png
 
-.. figure:: _static/SatohTestTemperature.png
+.. figure:: img/SatohTestTemperature.png
 
-.. figure:: _static/SatohTestPressure.png
+.. figure:: img/SatohTestPressure.png
 
 
 Compute Mean Thermodynamic Forces
@@ -74,7 +79,7 @@ We display the average stress SZZ over the RVE (composed of 83% matrix and 17% i
   awk '{if(NR>13) print $1 " " 0.83*$4+0.17*$10}' avgStress > res-mfem-mgis.txt
   plot "res-mfem-mgis.txt" u 1:2 w l title "mfem-mgis"
 
-.. figure:: _static/avgStress.png
+.. figure:: img/avgStress.png
 
 
 Compute Stored Energy
