+++ docs/source/user_guide/user_guide.rst
@@ -1,0 +1,12 @@
+.. _mfem_mgis_user_guide:
+
+==========
+User guide
+==========
+
+.. toctree::
+   :maxdepth: 1
+
+   tutorial.rst
+   post-processing.rst  
+   glossary.rst  

[…]
