+++ docs/source/user_guide/post-processing.rst
@@ -1,6 +1,8 @@
-========================
-Post Processing Features
-========================
+.. _mfem_mgis_post_processings:
+
+===============
+Post Processing
+===============
 
 In this section, we describe the post-processing options available in
 MFEM-MGIS.
