### 2025-03-19 7117d95
add a description of the available behaviour integrators
