+++ docs/source/user_guide/behaviour_integrators.rst
@@ -2,7 +2,7 @@
 Behaviour integrators
 =====================
 
-A Behaviour integrator compute residual and jacobian terms for a
+A Behaviour integrator computes residual and jacobian terms for a
 material. Behaviour integrators are associated with:
 
 - a physical phenomenon (mechanics, heat transfer, etc.),
@@ -15,7 +15,7 @@ material. Behaviour integrators are associated with:
 
 .. note::
 
-   Behaviour integrators are generally declared using the
+   Behaviour integrators are generally declared using 
    the :code:`addBehaviourIntegrator` method of the
    :code:`NonLinearEvolutionProblem` class, as in the
    following example:
@@ -36,16 +36,14 @@ Behaviour integrators and :code:`MFront` behaviours
 ===================================================
 
 Behaviour integrators rely on external behaviours that shall be
-generated by :code:`MFront`.
-
-In practice, those external behaviours are compiled by :code:`MFront`
+generated by :code:`MFront`. In practice, those external behaviours are compiled by :code:`MFront`
 into shared libraries that are loaded dynamically when an behaviour
 integrator is loaded.
 
 Mechanical behaviour integrator
 ===============================
 
-A mechanical behaviour integrator compute the residual and jacobian
+A mechanical behaviour integrator computes the residual and jacobian
 associated with one of the following variational operators:
 
 .. math::
@@ -57,7 +55,7 @@ associated with one of the following variational operators:
 
 where:
 
-- :math:`\Omega_r` is domain occupied by the material of interest in the
+- :math:`\Omega_r` is the domain occupied by the material of interest in the
   reference configuration,
 - :math:`\underline{\sigma}` is the `Cauchy` stress (expressed in the
   reference configuration for a small strain analysis,
@@ -76,7 +74,7 @@ support:
 Heat transfer behaviour integrators
 ===================================
 
-Nonlinear heat transfer behaviour integrator compute the residual and jacobian
+Nonlinear heat transfer behaviour integrators compute the residual and jacobian
 associated with the following variational operator:
 
 .. math::
@@ -85,7 +83,7 @@ associated with the following variational operator:
 
 where:
 
-- :math:`\Omega_r` is domain occupied by the material of interest in the
+- :math:`\Omega_r` is the domain occupied by the material of interest in the
   reference configuration,
 - :math:`\vec{j}` is the heat flux,
 - :math:`\vec{\nabla} T^{\star}` is the virtual temperature.
@@ -109,7 +107,7 @@ associated with the following variational operator:
 
 where:
 
-- :math:`\Omega_r` is domain occupied by the material of interest in the
+- :math:`\Omega_r` is the domain occupied by the material of interest in the
   reference configuration,
 - :math:`b_{\chi}` is the thermodynamic force associated with the
   gradient of the micromorphic damage :math:`\vec{\nabla} d_{\chi}`,
