+++ docs/source/developer_guide/developer_guide.rst
@@ -9,6 +9,6 @@ This project is written in ``C++-17``.
 .. toctree::
    :maxdepth: 1
 
-   benchmark.rst
    changelog.rst
+   benchmark.rst
    contributors.rst
