### 2025-03-19 72b2fa7
add a page about boundary conditions
