+++ docs/source/user_guide/post_processings.rst
@@ -15,8 +15,11 @@ MFEM-MGIS.
 
   This section is under construction
 
-Dump Paraview Results
-=====================
+The :code:`ParaviewExport` post-processing
+==========================================
+
+This post-processing allows to export the unknowns of a nonlinear
+evolution problem for visualization in :code:`paraview`:
 
 - Key: ``ParaviewExportResults``
 
