+++ include/CMakeLists.txt
@@ -29,6 +29,7 @@ mfem_mgis_header(MFEMMGIS DirichletBoundaryConditionBase.hxx)
 mfem_mgis_header(MFEMMGIS UniformDirichletBoundaryCondition.hxx)
 mfem_mgis_header(MFEMMGIS AbstractBoundaryCondition.hxx)
 mfem_mgis_header(MFEMMGIS UniformImposedPressureBoundaryCondition.hxx)
+mfem_mgis_header(MFEMMGIS UniformHeatSourceBoundaryCondition.hxx)
 mfem_mgis_header(MFEMMGIS ImposedDirichletBoundaryConditionAtClosestNode.hxx)
 mfem_mgis_header(MFEMMGIS PostProcessing.hxx)
 mfem_mgis_header(MFEMMGIS PostProcessingFactory.hxx)
