### 2025-03-24 714fa29
Update installation_guide.rst
