+++ cmake/modules/FindMFEM.cmake
@@ -108,6 +108,14 @@ if(MFEM_USE_MPI)
     include_directories(${MPI_INCLUDE_PATH})
 endif(MFEM_USE_MPI)
 
+# Optional MUMPS
+if(MFEM_USE_MUMPS)
+  find_package(OpenMP)
+  set (CMAKE_C_FLAGS "${CMAKE_C_FLAGS} ${OpenMP_C_FLAGS}")
+  set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
+  set (CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} ${OpenMP_EXE_LINKER_FLAGS}")
+endif(MFEM_USE_MUMPS)
+
 if (NOT MFEM_USE_SUITESPARSE)
   message(FATAL_ERROR "You shall have SUITESPARSE support activated in MFEM.CMake will exit.\n" )
 endif()
