+++ CMakeLists.txt
@@ -36,13 +36,6 @@ if (MFEM_USE_MPI)
   include_directories(SYSTEM ${METIS_INCLUDE_DIRS})
 endif(MFEM_USE_MPI)
 
-if (MFEM_USE_MUMPS)
-	find_package(OpenMP)
-  set (CMAKE_C_FLAGS "${CMAKE_C_FLAGS} ${OpenMP_C_FLAGS}")
-  set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
-  set (CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} ${OpenMP_EXE_LINKER_FLAGS}")
-endif (MFEM_USE_MUMPS)
-
 # MFontGenericInterface
 find_package (MFrontGenericInterface REQUIRED)
 
