### 2025-03-31 thelfer/mfem-mgis#102 @thelfer
@jacobmerson Could we close this issue ?
