### 2025-04-01 #7719 @iammix
Hi @thelfer,

Apologies for the delay in getting back to you, and thank you for the time and effort you and your team ve put into developing and maintaining this project. Your contribution to the open-source computational mechanics community is greatly appreciated.
my comments about the repo and the paper are listed below:

1. **Installation**: The installation steps are clearly documented and proceed as expected, though users must ensure that all dependencies (MFEM, MGIS, CMake ≥3.15) are installed in advance ✔️
2. **Functionality**: The examples (both C++ and Python) successfully demonstrate the software’s purpose and verify integration between MFEM and MGIS ✔️
3. **Performance**: No specific performance claims are made, and this is totally fine ✔️

4. **Statement of Need**: The README and paper clearly define the scientific need and target audience ✔️
5. **Installation Instructions**: Instructions are detailed, with alternatives provided. Dependencies are listed ✔️
6. **Usage Examples**: Provided examples show how the software works in real simulations ✔️
7. **API Documentation**: The documentation is linked in the README ✔️
8. **Testing**: There is a `tests` directory, but no clear instructions on how to run tests or verify installation. ❗Providing a brief explanation in the README or a `CONTRIBUTING.md` would be valuable.
9. **Community Guidelines**: Contribution and support guidelines are available ✔️

The paper is clear and well written, providing an accessible summary of the software’s purpose and its significance in computational mechanics. The "Statement of Need" section effectively explains the problems MFEM-MGIS addresses—namely, the coupling of a general finite element library (MFEM) with a behavior modeling infrastructure (MGIS) to simulate advanced material behaviors such as finite strain plasticity. The paper positions the software appropriately within the ecosystem of existing tools and justifies its contribution with well-cited references and context. The writing is concise and of high quality, and the references follow proper citation conventions.

While the submission is strong and ready for publication, I recommend the following **minor enhancements** which are up to you to fullfill them:

1. **Add brief test instructions** in the `README.md` or `CONTRIBUTING.md` (e.g., how to run unit tests or validate a build).
2. Optionally, provide a performance note or reference to help users benchmark it for their cases.

---

This is a solid and useful contribution to the open-source computational mechanics ecosystem. I recommend this software for publication in JOSS.
