### 2025-04-02 ec3ba9b
add contributing guide in the sphinx documentation
