### 2025-04-03 thelfer/mfem-mgis#103 @jacobmerson
It looks like you have added some additional notes on the usage in the web page. Between this and the examples, I am happy to close this issue.
