### 2025-04-03 #7719 @jacobmerson
@arfon @thelfer I am happy to recommend MFEM/MGIS for publication in JOSS. This is a very interesting and useful contribution to the community.
