### 2025-04-03 #7719 @thelfer
@iammix  Thanks for your feed-backs.

> Add brief test instructions in the README.md or CONTRIBUTING.md (e.g., how to run unit tests or validate a build).

@rprat-pro add a note in `CONTRIBUTING.md`

> Optionally, provide a performance note or reference to help users benchmark it for their cases.

A list of benchmarks are included here: https://thelfer.github.io/mfem-mgis/developer_guide/benchmark.html

For the tools, @rprat-pro has used the function declared in this header: https://github.com/thelfer/mfem-mgis/blob/master/include/MFEMMGIS/Profiler.hxx

Although the code is straighforward and an example of usage is provided in the `benchmarks` directory, he still has to document it properly and add a dedicated section in the documentation :)

@arfon On our side, everything is fine. Thanks again to the reviewers, thanks to them the paper and the documentation are much better than in the original version. This is exactly what we expect when submitting to JOSS !
