+++ docs/source/developer_guide/benchmark.rst
@@ -5,6 +5,96 @@ Benchmarks
     :depth: 3
     :local:
 
+Profiling tools
+^^^^^^^^^^^^^^^
+
+``MFEM-MGIS`` provides a simple tool for profiling your application, consisting mainly in generating a time table of timed sections respecting the call stack (tree). The master or root timer of the call stack is initialized during the general initialization of mfem-mgis: 
+
+.. code-block:: c++
+
+  mfem_mgis::initialize(argc, argv);
+
+Displaying the time table in your terminal is done by adding the following function call: 
+
+.. code-block:: c++
+
+  mfem_mgis::Profiler::timers::print_timers();
+
+This is an example of an output:
+
+.. code-block:: text
+
+  MPI feature activated, rank 0:
+  |-- start timetable -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+  |    name                                                                  |    number Of Calls |             min(s) |            mean(s) |             max(s) |            part(%) |             imb(%) |
+  |--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+  | > root                                                                   |                  1 |           0.061155 |           0.061313 |           0.061512 |        100.000000% |          0.003254% |
+  | |--> Solve                                                               |                  1 |           0.056359 |           0.056493 |           0.056539 |         91.914400% |          0.000812% |
+  |    |--> NonLinearEvolutionProblem::solve                                 |                  1 |           0.056359 |           0.056492 |           0.056538 |         91.913194% |          0.000812% |
+  |       |--> NonLinearEvolutionProblemImplementationBase::solve            |                  1 |           0.056357 |           0.056490 |           0.056536 |         91.909814% |          0.000810% |
+  |          |--> NewtonSolver::Mult                                         |                  1 |           0.056351 |           0.056483 |           0.056528 |         91.897665% |          0.000798% |
+  |             |--> NewtonSolver::computeNewtonCorrection                   |                 10 |           0.048653 |           0.048706 |           0.048851 |         79.416040% |          0.002969% |
+  |                |--> NewtonSolver::getJacobian                            |                 10 |           0.020212 |           0.020215 |           0.020220 |         32.870789% |          0.000239% |
+  |          |--> NonLinearEvolutionProblemImplementationBase::setup         |                  1 |           0.000004 |           0.000005 |           0.000006 |          0.009562% |          0.148211% |
+  | |--> FiniteElementDiscretization::constructor                            |                  1 |           0.002696 |           0.002717 |           0.002760 |          4.487192% |          0.015873% |
+  |    |--> loadMesh                                                         |                  1 |           0.000944 |           0.000965 |           0.001002 |          1.629139% |          0.038049% |
+  | |--> common::post_processing_step                                        |                  1 |           0.000856 |           0.000974 |           0.001084 |          1.762192% |          0.113317% |
+  | |--> set_mgis_stuff                                                      |                  1 |           0.000687 |           0.000824 |           0.000970 |          1.577426% |          0.177846% |
+  | |--> set_linear_solver                                                   |                  1 |           0.000139 |           0.000169 |           0.000204 |          0.331618% |          0.205352% |
[…]
