### 2025-04-03 1ec4819
Update benchmark.rst typo
