### 2025-04-07 #7719 @rprat-pro
@arfon - The new release `v1.0.2` has been uploaded on Zenodo with the latest recommendations from reviewers at https://zenodo.org/records/15166161 .
