### 2025-04-10 #7719 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
